requirements.txt
# Core quantum libraries
qiskit==1.0.0
qiskit-aer==0.13.0
pennylane==0.35.0
pennylane-qiskit==0.35.0

# ML and embeddings - optimized for quantum-inspired similarity
torch>=2.6.0
sentence-transformers>=2.2.2
transformers>=4.36.0
faiss-cpu>=1.7.4
numpy>=1.24.3

# API and utilities
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
loguru>=0.7.2
gunicorn>=21.2.0

# Configuration and data handling
pyyaml>=6.0.1
toml>=0.10.2
python-multipart>=0.0.6
aiofiles>=23.2.1

# Security and authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
cryptography>=41.0.8

# Caching and performance
redis>=5.0.1
aioredis>=2.0.1
cachetools>=5.3.2

# Monitoring and metrics
prometheus-client>=0.19.0
psutil>=5.9.6

# Additional dependencies for quantum-inspired embeddings
scipy>=1.11.4
scikit-learn>=1.3.2
pandas>=2.1.4
matplotlib>=3.8.2

# HTTP client for health checks
httpx>=0.25.2
requests>=2.31.0